Fix clang dependencies and minor cleanup fixes.
- Clang was picking up libzstd from the host environment which
was causing the shiboken2 build to fail on the macos_arm_default
profile. This change adds zstd as a conan dependency so the
problematic host environment version is not used.

- Made a few other minor qt & libraw option changes to resolve
some issues uncovered during the investigation of the
zstd issue. These were discovered when trying to build with
conan's new tools.cmake.cmakedeps:new=will_break_next config
option.

- Removed old llvm recipe that is not used anymore.